Add search results page, refactor landingpage
