+++ client/src/components/App.js
@@ -16,6 +16,7 @@ import CartPage from "./views/CartPage/CartPage";
 import BeatPage from "./views/BeatPage/BeatPage";
 import BeatPlayer from "./views/BeatPlayer/BeatPlayer";
 import DownloadPage from "./views/DownloadPage/DownloadPage";
+import SearchResultsPage from './views/SearchResultsPage/SearchResultsPage';
 import ErrorNotFoundPage from "./views/ErrorNotFoundPage/ErrorNotFoundPage";
 import { AudioContextProvider } from "./utils/AudioContext";
 import { ThemeProvider, CSSReset } from "@chakra-ui/core";
@@ -74,6 +75,7 @@ function App() {
             <Route exact path="/cart" component={Auth(CartPage, PUBLIC_PAGE, false)} />
             <Route exact path="/orders" component={Auth(OrdersPage, LOGGED_IN_ONLY, false)} />
             <Route exact path="/download" component={Auth(DownloadPage, PUBLIC_PAGE, false)} />
+            <Route exact path="/beats" component={Auth(SearchResultsPage, PUBLIC_PAGE, false)} />
             <Route component={Auth(ErrorNotFoundPage, PUBLIC_PAGE, false)} />
           </Switch>
         </NavBar>
